TIP-882: move datagrid services
