+++ src/Oro/Bundle/PimFilterBundle/Resources/config/filters.yml
@@ -201,3 +201,14 @@ services:
             - '@pim_filter.product_utility'
         tags:
             - { name: oro_filter.extension.orm_filter.filter, type: label_or_identifier }
+
+    pim_reference_data.filter.reference_data:
+        class: 'Oro\Bundle\PimFilterBundle\Filter\ProductValue\ReferenceDataFilter'
+        arguments:
+            - '@form.factory'
+            - '@pim_filter.product_utility'
+            - '@pim_user.context.user'
+            - '@pim_catalog.repository.attribute'
+            - '@pim_reference_data.registry'
+        tags:
+            - { name: oro_filter.extension.orm_filter.filter, type: product_value_reference_data }
