--- src/Pim/Bundle/ReferenceDataBundle/Resources/config/datagrid/normalizers.yml
@@ -1,14 +1,0 @@
-parameters:
-    pim_reference_data.datagrid.normalizer.reference_data.class: Pim\Bundle\ReferenceDataBundle\DataGrid\Normalizer\ReferenceDataNormalizer
-    pim_reference_data.datagrid.normalizer.reference_data_collection.class: Pim\Bundle\ReferenceDataBundle\DataGrid\Normalizer\ReferenceDataCollectionNormalizer
-
-services:
-    pim_reference_data.datagrid.normalizer.reference_data:
-        class: '%pim_reference_data.datagrid.normalizer.reference_data.class%'
-        tags:
-            - { name: pim_datagrid_serializer.normalizer }
-
-    pim_reference_data.datagrid.normalizer.reference_data_collection:
-        class: '%pim_reference_data.datagrid.normalizer.reference_data_collection.class%'
-        tags:
-            - { name: pim_datagrid_serializer.normalizer }
