--- src/Pim/Bundle/ReferenceDataBundle/Resources/config/datagrid/sorters.yml
@@ -1,10 +1,0 @@
-parameters:
-    pim_reference_data.extension.sorter.reference_data_sorter.class: Pim\Bundle\ReferenceDataBundle\DataGrid\Extension\Sorter\ReferenceDataSorter
-
-services:
-    pim_reference_data.extension.sorter.reference_data_sorter:
-        class: '%pim_reference_data.extension.sorter.reference_data_sorter.class%'
-        arguments:
-            - '@oro_datagrid.datagrid.request_params'
-        tags:
-            - { name: pim_datagrid.extension.sorter, type: product_value_reference_data }
